+++ CardinalKit-Example/CardinalKit/Extensions/Notification.swift
@@ -1,0 +1,14 @@
+//
+//  Notification.swift
+//  CardinalKit_Example
+//
+//  Created by Esteban Ramos on 24/06/22.
+//  Copyright © 2022 CocoaPods. All rights reserved.
+//
+
+import Foundation
+
+extension Notification.Name {
+    static let onBoardingStateChange = Notification.Name("onBoardingStateChange")
+    static let onUserStateChange = Notification.Name("onUserStateChange")
+}
