+++ CardinalKit-Example/CardinalKit/Presenters/LaunchPresenter.swift
@@ -1,0 +1,36 @@
+//
+//  LaunchPresenter.swift
+//  CardinalKit_Example
+//
+//  Created by Esteban Ramos on 24/06/22.
+//  Copyright © 2022 CocoaPods. All rights reserved.
+//
+
+import Foundation
+import SwiftUI
+
+class LaunchPresenter:ObservableObject{
+    @Published var didCompleteOnBoarding:Bool
+    
+    init(){
+        didCompleteOnBoarding = UserDefaults.standard.bool(forKey: Constants.onboardingDidComplete) && Libraries.shared.authlibrary.user != nil
+        
+        NotificationCenter.default.addObserver(self, selector: #selector(onBoardingStateChange), name: .onBoardingStateChange, object: nil)
+        
+        NotificationCenter.default.addObserver(self, selector: #selector(onUserStateChange), name: .onUserStateChange, object: nil)
+    }
+    
+    @objc func onBoardingStateChange(_ notification: Notification){
+        if let newValue = notification.object as? Bool {
+            didCompleteOnBoarding = newValue
+        }
+    }
+    
+    @objc func onUserStateChange(_ notification: Notification){
+        if let newValue = notification.object as? Bool,
+            !newValue {
+            didCompleteOnBoarding = false
+            UserDefaults.standard.set(false, forKey: Constants.onboardingDidComplete)
+        }
+    }
+}
