+++ CardinalKit-Example/CardinalKit/Views/OnBoarding/CommonSteps/CKReviewConsentDocument.swift
@@ -8,7 +8,6 @@
 
 import Foundation
 import ResearchKit
-import Firebase
 
 
 public class CKReviewConsentDocument: ORKQuestionStep {
@@ -31,28 +30,19 @@ public class CKReviewConsentDocumentViewController:ORKQuestionStepViewController
     }
     
     public override func viewDidLoad() {
-        let storage = Storage.storage()
-        let storageRef = storage.reference()
-        
         // Check if a consent document exists on the cloud, otherwise user will need to re-consent
         if let DocumentCollection = CKStudyUser.shared.authCollection {
             let config = CKPropertyReader(file: "CKConfiguration")
             let consentFileName = config.read(query: "Consent File Name")
-            let DocumentRef = storageRef.child("\(DocumentCollection)/\(consentFileName).pdf")
-            
             var docURL = (FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)).last as NSURL?
             docURL = docURL?.appendingPathComponent("\(consentFileName).pdf") as NSURL?
             let url = docURL! as URL
             
-            DocumentRef.write(toFile: url) { url, error in
-                if let error = error {
-                    print(error.localizedDescription)
-                    self.setAnswer(false)
-                } else {
-                    self.setAnswer(true)
-                }
+            Libraries.shared.networkingLibrary.checkIfFileExist(url: url, path: "\(DocumentCollection)/\(consentFileName).pdf", onComplete: {
+                exist in
+                self.setAnswer(exist)
                 super.goForward()
-            }
+            })
         }
     }
 }
